Bugfix for NEMO-like land processes elimination.
Values describing a domain (latitude, longitude, bounds,...) are set to -1 on land processes.


git-svn-id: https://forge.ipsl.jussieu.fr/ioserver/svn/XIOS/branchs/xios-2.5@1624 d613d254-d18a-4304-b296-5125d4c4d670

+++ src/node/axis.hpp
@@ -123,8 +123,8 @@ namespace xios {
          bool zoomByIndex();
 
       public: 
-        bool hasValue;        
-        CArray<size_t,1> localIndexToWriteOnServer;        
+        bool hasValue;
+        CArray<int,1> localIndexToWriteOnServer;
 
       private:
          void checkData();
